# api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@trexis.net
    name: treXis LLC
    url: https://www.trexis.net
  description: The public facing API through which connectors are exposed as a single
    abstract API
  termsOfService: http://www.trexis.net/termsofservices
  title: Cloud API
  version: v1.5
servers:
- url: /
security:
- cloudAPI: []
- connectorToken: []
paths:
  /cloud/v1/entity/{entityId}:
    get:
      description: Return a given entity's information
      operationId: getEntity
      parameters:
      - description: Unique Entity Identifier
        explode: false
        in: path
        name: entityId
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates if data in the response should be masked
        explode: true
        in: query
        name: mask
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Indicates if the request should be processed for enhancement
          after getting data from the connector
        explode: true
        in: query
        name: enhance
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Value used for tracing requests and responses
        explode: false
        in: header
        name: X-TRACE-ID
        required: false
        schema:
          example: "1234"
          type: string
        style: simple
      - description: Token to carry/send information between requests. Connectors
          generate and validate the token content and type.
        explode: false
        in: header
        name: X-TOKEN
        required: false
        schema:
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
          description: Entity Model
          headers:
            X-TRACE-ID:
              $ref: '#/components/headers/X-TRACE-ID'
            X-TOKEN:
              $ref: '#/components/headers/X-TOKEN'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Return entity by entityId
      tags:
      - Entity
  /cloud/v1/entity/{entityId}/profile:
    get:
      description: Return a given entity's profile information
      operationId: getEntityProfile
      parameters:
      - description: Unique Entity Identifier
        explode: false
        in: path
        name: entityId
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates if data in the response should be masked
        explode: true
        in: query
        name: mask
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Indicates if the request should be processed for enhancement
          after getting data from the connector
        explode: true
        in: query
        name: enhance
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Value used for tracing requests and responses
        explode: false
        in: header
        name: X-TRACE-ID
        required: false
        schema:
          example: "1234"
          type: string
        style: simple
      - description: Token to carry/send information between requests. Connectors
          generate and validate the token content and type.
        explode: false
        in: header
        name: X-TOKEN
        required: false
        schema:
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityProfile'
          description: Entity Profile
          headers:
            X-TRACE-ID:
              $ref: '#/components/headers/X-TRACE-ID'
            X-TOKEN:
              $ref: '#/components/headers/X-TOKEN'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Return entity profile
      tags:
      - Entity
    put:
      description: Update a given entity's profile information
      operationId: putEntityProfile
      parameters:
      - description: Unique Entity Identifier
        explode: false
        in: path
        name: entityId
        required: true
        schema:
          type: string
        style: simple
      - description: Indicates if data in the response should be masked
        explode: true
        in: query
        name: mask
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Indicates if the request should be processed for enhancement
          after getting data from the connector
        explode: true
        in: query
        name: enhance
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Value used for tracing requests and responses
        explode: false
        in: header
        name: X-TRACE-ID
        required: false
        schema:
          example: "1234"
          type: string
        style: simple
      - description: Token to carry/send information between requests. Connectors
          generate and validate the token content and type.
        explode: false
        in: header
        name: X-TOKEN
        required: false
        schema:
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityProfile'
        description: Profile to update
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityProfile'
          description: Entity Profile
          headers:
            X-TRACE-ID:
              $ref: '#/components/headers/X-TRACE-ID'
            X-TOKEN:
              $ref: '#/components/headers/X-TOKEN'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Update entity profile
      tags:
      - Entity
  /cloud/v1/entities:
    post:
      description: Create a new entity based on the given profile information
      operationId: postEntityProfile
      parameters:
      - description: Indicates if data in the response should be masked
        explode: true
        in: query
        name: mask
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Indicates if the request should be processed for enhancement
          after getting data from the connector
        explode: true
        in: query
        name: enhance
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Value used for tracing requests and responses
        explode: false
        in: header
        name: X-TRACE-ID
        required: false
        schema:
          example: "1234"
          type: string
        style: simple
      - description: Token to carry/send information between requests. Connectors
          generate and validate the token content and type.
        explode: false
        in: header
        name: X-TOKEN
        required: false
        schema:
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityProfile'
        description: Profile to create
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
          description: Entity Profile
          headers:
            X-TRACE-ID:
              $ref: '#/components/headers/X-TRACE-ID'
            X-TOKEN:
              $ref: '#/components/headers/X-TOKEN'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Create entity
      tags:
      - Entity
  /cloud/v1/entities/search:
    post:
      description: Return a paginated list of entities based on a search filter
      operationId: searchEntities
      parameters:
      - description: Number of entries to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      - description: Base64 encoded information on how to retrieve the next set of
          results within the cursor.
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          example: YWNjb3VudE51bWJlcjoxMjM0MTIzNCxhY2NvdW50VHlwZT1EREE=
          type: string
        style: form
      - description: Indicates if data in the response should be masked
        explode: true
        in: query
        name: mask
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Indicates if the request should be processed for enhancement
          after getting data from the connector
        explode: true
        in: query
        name: enhance
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: Value used for tracing requests and responses
        explode: false
        in: header
        name: X-TRACE-ID
        required: false
        schema:
          example: "1234"
          type: string
        style: simple
      - description: Token to carry/send information between requests. Connectors
          generate and validate the token content and type.
        explode: false
        in: header
        name: X-TOKEN
        required: false
        schema:
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/SearchFilter'
              type: array
        description: Search for information based on name value pairs with defined
          opperators
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityList'
          description: Entity List
          headers:
            X-TRACE-ID:
              $ref: '#/components/headers/X-TRACE-ID'
            X-TOKEN:
              $ref: '#/components/headers/X-TOKEN'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
      summary: Search for entities
      tags:
      - Entity
...
components:
  headers:
    X-TRACE-ID:
      description: Value used for tracing requests and responses
      explode: false
      required: false
      schema:
        example: "1234"
        type: string
      style: simple
    X-TOKEN:
      description: Token to carry/send information between requests. Connectors generate
        and validate the token content and type.
      explode: false
      required: false
      schema:
        example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
        type: string
      style: simple
...
  schemas:
    Entity:
      description: Entity Object
      example:
        dateCreated: YYYY-MM-DDThh:mm:ssZ
        name: BridgeLine Inc or Luke Skywalker
        active: true
        securityInformation:
          lastLoginDate: YYYY-MM-DDThh:mm:ss
          username: user01
          token: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
        id: id
        taxInformation:
        - taxId: taxId
          type: SSN
        - taxId: taxId
          type: SSN
        type: PERSON
        URI:
          URL: http://host/images/1234.jpg
      properties:
        type:
          description: The type of profile
          enum:
          - PERSON
          - BUSINESS
          - SYSTEM
          type: string
        id:
          description: Unique identifier of the entity
          type: string
        name:
          description: Name of the entity
          example: BridgeLine Inc or Luke Skywalker
          type: string
        taxInformation:
          description: Entity tax information
          items:
            $ref: '#/components/schemas/TaxInformation'
          type: array
        active:
          description: Indicates whether entity is active
          type: boolean
        dateCreated:
          description: ISO 8601 Date entity got created
          example: YYYY-MM-DDThh:mm:ssZ
          type: string
        securityInformation:
          $ref: '#/components/schemas/SecurityInformation'
        URI:
          $ref: '#/components/schemas/FiniteURI'
      required:
      - id
      type: object
    EntityProfile:
      description: Entity Profile Information
      example:
        firstname: Luke
        addresses:
        - country: United States of America
          city: Alexander City
          postalCode: "99999"
          state: Alabama
          type: BUSINESS
          line3: Sub Department
          line2: Department or House Number
          line1: Street Name or PO Box
          primary: true
        - country: United States of America
          city: Alexander City
          postalCode: "99999"
          state: Alabama
          type: BUSINESS
          line3: Sub Department
          line2: Department or House Number
          line1: Street Name or PO Box
          primary: true
        preferences:
        - name: Language
          value: English
        - name: Language
          value: English
        contactPoints:
        - prefix: "1"
          name: HOMEPHONE or WORKPHONE
          type: SMS
          suffix: ext 123
          value: 555-555-5555
        - prefix: "1"
          name: HOMEPHONE or WORKPHONE
          type: SMS
          suffix: ext 123
          value: 555-555-5555
        dateOfBirth: yyyy-MM-dd
        id: id
        fullname: BridgeLine Inc
        taxInformation:
        - taxId: taxId
          type: SSN
        - taxId: taxId
          type: SSN
        type: PERSON
        URI:
          URL: http://host/images/1234.jpg
        lastname: Skywalker
      properties:
        type:
          description: The type of profile
          enum:
          - PERSON
          - BUSINESS
          - SYSTEM
          type: string
        id:
          description: Unique identifier of the entity
          type: string
        fullname:
          description: The fullname of a the entity
          example: BridgeLine Inc
          type: string
        firstname:
          description: The legal firstname of the person
          example: Luke
          type: string
        lastname:
          description: The legal lastname/surname of the person
          example: Skywalker
          type: string
        dateOfBirth:
          description: ISO 8601 Date of Birth
          example: yyyy-MM-dd
          type: string
        addresses:
          description: Entity addresses
          items:
            $ref: '#/components/schemas/Address'
          type: array
        contactPoints:
          description: Entity contact points
          items:
            $ref: '#/components/schemas/ContactPoint'
          type: array
        preferences:
          description: Entity preferences
          items:
            $ref: '#/components/schemas/Attribute'
          type: array
        taxInformation:
          description: Entity tax information
          items:
            $ref: '#/components/schemas/TaxInformation'
          type: array
        URI:
          $ref: '#/components/schemas/FiniteURI'
      required:
      - id
      type: object
...
    Address:
      description: Address object
      example:
        country: United States of America
        city: Alexander City
        postalCode: "99999"
        state: Alabama
        type: BUSINESS
        line3: Sub Department
        line2: Department or House Number
        line1: Street Name or PO Box
        primary: true
      properties:
        type:
          enum:
          - BUSINESS
          - RESIDENTIAL
          - POBOX
          type: string
        line1:
          description: Address line
          example: Street Name or PO Box
          type: string
        line2:
          description: Address line
          example: Department or House Number
          type: string
        line3:
          description: Address line
          example: Sub Department
          type: string
        city:
          description: Address City or Town
          example: Alexander City
          type: string
        state:
          description: State or Region
          example: Alabama
          type: string
        postalCode:
          description: Postal code or Zip code
          example: "99999"
          type: string
        country:
          description: Country
          example: United States of America
          type: string
        primary:
          description: Primary Address
          type: boolean
      type: object
    ContactPoint:
      description: Contact Point object
      example:
        prefix: "1"
        name: HOMEPHONE or WORKPHONE
        type: SMS
        suffix: ext 123
        value: 555-555-5555
      properties:
        type:
          enum:
          - SMS
          - EMAIL
          - VOICE
          type: string
        name:
          description: Name of the contact point
          example: HOMEPHONE or WORKPHONE
          type: string
        prefix:
          description: Text preceeding contact point
          example: "1"
          type: string
        value:
          description: Contact point value
          example: 555-555-5555
          type: string
        suffix:
          description: Additional information
          example: ext 123
          type: string
      type: object
    Attribute:
      description: Name value pair property
      example:
        name: Language
        value: English
      properties:
        name:
          description: Attribute name
          example: Language
          type: string
        value:
          description: Attribute value
          example: English
          type: string
      type: object
    TaxInformation:
      description: Tax information
      example:
        taxId: taxId
        type: SSN
      properties:
        type:
          enum:
          - SSN
          - EIN
          - ITIN
          - ATIN
          - OTHER
          type: string
        taxId:
          description: Tax identification value
          type: string
      type: object
    SecurityInformation:
      description: Security Information
      example:
        lastLoginDate: YYYY-MM-DDThh:mm:ss
        username: user01
        token: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
      properties:
        username:
          description: Unique username
          example: user01
          type: string
        token:
          description: Unique token to identify entity for subsequent requests.  E.g.
            OAuth access_token
          example: VXNlclNlc3Npb25Ub2tlbjoxMjM0MTIzNA==
          type: string
        lastLoginDate:
          description: ISO 8601 date the entity last logged in
          example: YYYY-MM-DDThh:mm:ss
          type: string
      type: object
    SearchFilter:
      description: Search filter entry
      example:
        field: username
        query: lskywalker
        operator: EQUAL
      properties:
        field:
          description: The name of the field of element to query
          example: username
          type: string
        query:
          description: The value to query for
          example: lskywalker
          type: string
        operator:
          enum:
          - EQUAL
          - LIKE
          - GREATER
          - LESS
          - NOT_LIKE
          - NOT_EQUAL
          - IN
          - NOT_IN
          type: string
      type: object
    FiniteURI:
      description: Unique Resource Identifier in the API
      example:
        URL: http://host/images/1234.jpg
      properties:
        URL:
          description: The URL to the URI
          example: http://host/images/1234.jpg
          type: string
      type: object
...
    BaseListProperties:
      description: Base properties for all lists object
      properties:
        status:
          description: Indicates whether a operation was successful
          type: boolean
        messages:
          description: Potential information or errror messages
          items:
            description: Potential information or error messages
            type: string
          type: array
        totalItems:
          description: The total number of list items that exist for the list or query
          example: 200
          type: integer
        nextCursor:
          description: The Base64 encoded value that indicates the next cursor. The
            cursor will be decoded to use as the starting point to fetch the next
            set of results.
          example: YWNjb3VudE51bWJlcjoxMjM0MTIzNCxhY2NvdW50VHlwZT1EREE=
          type: string
      type: object
    EntityList:
      allOf:
      - $ref: '#/components/schemas/BaseListProperties'
      - $ref: '#/components/schemas/EntityList_allOf'
...
    BaseErrorProperties:
      description: Base properties for all lists object
      properties:
        status:
          description: Indicates whether a operation was successful
          type: boolean
        messages:
          description: Potential information or errror messages
          items:
            description: Potential information or error messages
            type: string
          type: array
      type: object
    ErrorResponse:
      allOf:
      - $ref: '#/components/schemas/BaseErrorProperties'
...
    EntityList_allOf:
      properties:
        entities:
          description: The list of entities
          items:
            $ref: '#/components/schemas/Entity'
          type: array
...
  securitySchemes:
    cloudAPI:
      description: The unique tenant API key allowing access to the API and connectors
      in: header
      name: X-API-KEY
      type: apiKey
    connectorToken:
      description: The unique key that can be passed to the connector
      in: header
      name: X-TOKEN
      type: apiKey